requirements.txt
asgiref==3.8.1
Django==5.2.4
django-browser-reload==1.13.0
django-cors-headers==4.7.0
djangorestframework==3.16.1
numpy==2.3.1
pandas==2.3.0
pillow==11.3.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
